requirements.txt
python-magic
argparse
